scripts/requirements.txt
requests>=2.31.0
openai>=1.51.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
cryptography>=41.0.0
pillow>=10.0.0